src/requirements.txt
aiohttp
SQLAlchemy
asyncpg
alembic
APScheduler
python-dotenv